repository_spec and repository_spec_parse untangling
ansible_galaxy.repository_spec and ansible_galaxy.repository_spec_parse
had lots of duplicated code between them. There were various reasons
for that, but none of them good.

With lots of role/repository related code removed, unduplicating
them got easier, so now they are undup'ed.
